requirements.txt
Pillow
requests
python-dotenv
google-generativeai
playsound
pygame

